# openapi.yaml
openapi: "3.1.0"
info:
  title: UIDE Inscripciones API
  version: "2.0.0"
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - ApiKeyAuth: []
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Estado del servidor
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Servidor activo.
          content:
            application/json:
              example:
                status: ok
                ts: "2026-06-23T10:00:00.000Z"
        '401':
          description: API key ausente o inválida.
          content:
            application/json:
              example:
                error: API key inválida o ausente
  /v1/inscripciones:
    post:
      operationId: crearInscripcionV1
      summary: Crear inscripción (contrato v1)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscripcionV1'
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440000"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"
      responses:
        '201':
          description: Inscripción creada.
          content:
            application/json:
              example:
                version: v1
                estudianteId: "550e8400-e29b-41d4-a716-446655440000"
                materias:
                  - "LTI_05A_458"
                periodoId: "2026-1"
        '400':
          description: Payload inválido.
          content:
            application/json:
              example:
                error: "Campos requeridos: estudianteId, materias, periodoId"
        '401':
          description: API key ausente o inválida.
  /v2/inscripciones:
    post:
      operationId: crearInscripcionV2
      summary: Crear inscripción con método de pago (contrato v2)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InscripcionV2'
            example:
              estudianteId: "550e8400-e29b-41d4-a716-446655440001"
              materias:
                - "LTI_05A_458"
              periodoId: "2026-1"
              payment_method: scholarship
      responses:
        '201':
          description: Inscripción creada.
          content:
            application/json:
              example:
                version: v2
                estudianteId: "550e8400-e29b-41d4-a716-446655440001"
                materias:
                  - "LTI_05A_458"
                periodoId: "2026-1"
                payment_method: scholarship
        '400':
          description: Payload inválido o payment_method fuera de enum.
          content:
            application/json:
              examples:
                campos_faltantes:
                  value:
                    error: "Campos requeridos: estudianteId, materias, periodoId, payment_method"
                metodo_invalido:
                  value:
                    error: "payment_method invalido. Valores: debit, credit, scholarship"
        '401':
          description: API key ausente o inválida.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    InscripcionV1:
      type: object
      required:
        - estudianteId
        - materias
        - periodoId
      properties:
        estudianteId:
          type: string
          format: uuid
        materias:
          type: array
          items:
            type: string
          minItems: 1
        periodoId:
          type: string
    InscripcionV2:
      allOf:
        - $ref: '#/components/schemas/InscripcionV1'
        - type: object
          required:
            - payment_method
          properties:
            payment_method:
              type: string
              enum:
                - debit
                - credit
                - scholarship
      unevaluatedProperties: false